events/requirements.txt
Django==4.2.12
djangorestframework==3.16.0
psycopg2-binary==2.9.10  # PostgreSQL adapter
django-cors-headers==4.7.0 #access backend
pika==1.3.2 #for rabbitMQ
# Add to requirements.txt
django-filter==25.1  # Use the appropriate version